Created a bunch of APIs
Created all the APIs that I could think of, we can add more as we go and find out we need something. Note that there are not delete apis because Im actually not sure how that works, but I'll look into it. Also creating a reservation correctly updates the number of reservations for that Event, but that was the last thing I added, so the numbers might be wrong for some events I was testing before adding it (we'll prob just clean up the whole database eventually anyways.)